Prevent Emacs daemon from crashing at logging out the GUI session.
Emacs daemon dies when there has been an X connection and the X
server is suddenly down for some reason even if there is no frame
in X anymore (see bug#11741).  This is because X connection remains
after the last frame in X is deleted.  So, explicitly closing the
connection at the frame deletion solves this problem.

Emacs daemon should not have SESSION_MANAGER or otherwise the
session manager may ask the daemon to exit when the session ends.